move to parent for logic and template only component
