+++ ui/app/adapters/kv/metadata.js
@@ -69,7 +69,7 @@ export default class KvMetadataAdapter extends ApplicationAdapter {
   }
 
   deleteRecord(store, type, snapshot) {
-    const { backend, fullSecretPath } = snapshot.record;
-    return this.ajax(this._url(kvMetadataPath(backend, fullSecretPath)), 'DELETE');
+    const { backend, path } = snapshot.record;
+    return this.ajax(this._url(kvMetadataPath(backend, path)), 'DELETE');
   }
 }
