+++ ui/lib/kv/addon/components/kv-delete-modal.hbs
@@ -1,61 +1,32 @@
-<button type="button" class="toolbar-link" {{on "click" (fn this.handleButtonClick @mode)}} data-test-kv-delete={{@mode}}>
-  {{capitalize @mode}}
+<button type="button" class="toolbar-link" {{on "click" @toggleModal}} data-test-kv-delete={{@mode}}>
+  {{@buttonText}}
 </button>
-{{#if this.modalOpen}}
+{{#if @modalOpen}}
   <Modal
-    @title="{{capitalize @mode}} {{if (not-eq @mode 'delete metadata') 'version?' ''}}"
-    @onClose={{action (mut this.modalOpen) false}}
-    @isActive={{this.modalOpen}}
-    @type={{if (eq @mode "delete") "warning" "danger"}}
+    @title="{{@buttonText}} {{if (not-eq @mode 'delete-metadata') 'version?' ''}}"
+    @onClose={{@toggleModal}}
+    @isActive={{@toggleModal}}
+    @type={{if (eq @buttonText "Delete") "warning" "danger"}}
     @showCloseButton={{true}}
   >
     <section class="modal-card-body">
       <p class="has-bottom-margin-s">
-        {{this.modalIntro}}
+        {{yield to="description"}}
       </p>
-      {{#if (eq @mode "delete")}}
-        <div class="is-flex-column">
-          {{#each this.generateDeleteRadioOptions as |option|}}
-            <ToolTip @verticalPosition="above" @horizontalPosition="left" as |T|>
-              <T.Trigger @tabindex="-1">
-                <div class="is-flex-align-baseline has-bottom-margin-m">
-                  <RadioButton
-                    id={{dasherize option.label}}
-                    class="radio top-xxs"
-                    @disabled={{option.disabled}}
-                    @value={{option.key}}
-                    @groupValue={{this.deleteType}}
-                    @onChange={{fn (mut this.deleteType) option.key}}
-                  />
-                  <div class="has-left-margin-s {{if option.disabled 'text-disabled'}}">
-                    <label class="has-text-weight-semibold">{{option.label}}</label>
-                    <p>{{option.description}}</p>
-                  </div>
-                </div>
-              </T.Trigger>
-              {{#if option.disabled}}
-                <T.Content @defaultClass="tool-tip">
-                  <div class="box">
-                    {{option.tooltipMessage}}
-                  </div>
-                </T.Content>
-              {{/if}}
-            </ToolTip>
-          {{/each}}
-        </div>
-      {{/if}}
+
+      {{yield to="content"}}
+
     </section>
     <footer class="modal-card-foot modal-card-foot-outlined">
       <button
         type="button"
-        class="button
-          {{if (or (eq @mode 'destroy') (eq @mode 'delete metadata')) 'is-danger-outlined' 'is-warning-outlined'}}"
-        {{on "click" (perform this.save)}}
-        disabled={{(not this.deleteType)}}
+        class="button {{if (eq @mode 'delete') 'is-warning-outlined' 'is-danger-outlined'}}"
+        {{on "click" @handleDelete}}
+        disabled={{@disabled}}
       >
-        {{capitalize @mode}}
+        {{capitalize @buttonText}}
       </button>
-      <button type="button" class="button is-secondary" {{on "click" (fn (mut this.modalOpen) false)}}>
+      <button type="button" class="button is-secondary" {{on "click" @toggleModal}}>
         Cancel
       </button>
     </footer>
