+++ ui/lib/kv/addon/components/page/secret/details.hbs
@@ -20,15 +20,77 @@
   <:toolbarActions>
     {{! Start of delete options }}
     {{#if (and @secret.deletionTime @secret.canUndelete (not @secret.destroyed))}}
-      <KvDeleteModal @mode="undelete" @path={{@secret.path}} @secret={{@secret}} />
+      <button
+        type="button"
+        class="toolbar-link"
+        {{on "click" (fn this.handleDelete "undelete")}}
+        data-test-kv-delete="undelete-version"
+      >
+        Undelete
+      </button>
     {{/if}}
     {{#if
       (and (not @secret.deletionTime) (not @secret.destroyed) (or @secret.canDeleteVersion @secret.canDeleteLatestVersion))
     }}
-      <KvDeleteModal @mode="delete" @path={{@secret.path}} @secret={{@secret}} @metadata={{@metadata}} />
+      <KvDeleteModal
+        @mode="delete"
+        @toggleModal={{(fn this.toggleModal "delete")}}
+        @modalOpen={{this.deleteModalOpen}}
+        @handleDelete={{(fn this.handleDelete "delete")}}
+        @buttonText="Delete"
+        @disabled={{(not this.deleteType)}}
+      >
+        <:description>
+          There are two ways to delete a version of a secret. Both delete actions can be undeleted later. How would you like
+          to proceed?
+        </:description>
+        <:content>
+          <div class="is-flex-column">
+            {{#each this.generateDeleteRadioOptions as |option|}}
+              <ToolTip @verticalPosition="above" @horizontalPosition="left" as |T|>
+                <T.Trigger @tabindex="-1">
+                  <div class="is-flex-align-baseline has-bottom-margin-m">
+                    <RadioButton
+                      id={{dasherize option.label}}
+                      class="radio top-xxs"
+                      @disabled={{option.disabled}}
+                      @value={{option.key}}
+                      @groupValue={{this.deleteType}}
+                      @onChange={{fn (mut this.deleteType) option.key}}
+                    />
+                    <div class="has-left-margin-s {{if option.disabled 'text-disabled'}}">
+                      <label class="has-text-weight-semibold">{{option.label}}</label>
+                      <p>{{option.description}}</p>
+                    </div>
+                  </div>
+                </T.Trigger>
+                {{#if option.disabled}}
+                  <T.Content @defaultClass="tool-tip">
+                    <div class="box">
+                      {{option.tooltipMessage}}
+                    </div>
+                  </T.Content>
+                {{/if}}
+              </ToolTip>
+            {{/each}}
+          </div>
+        </:content>
+      </KvDeleteModal>
     {{/if}}
     {{#if (and (not @secret.destroyed) @secret.canDestroyVersion)}}
-      <KvDeleteModal @mode="destroy" @path={{@secret.path}} @secret={{@secret}} />
+      <KvDeleteModal
+        @mode="destroy-version"
+        @toggleModal={{(fn this.toggleModal "destroy")}}
+        @modalOpen={{this.destroyModalOpen}}
+        @handleDelete={{(fn this.handleDelete "destroy")}}
+        @buttonText="Destroy"
+      >
+        <:description>
+          This action will permanently destroy Version
+          {{@secret.version}}
+          of the secret, and the secret data cannot be read or recovered later.
+        </:description>
+      </KvDeleteModal>
     {{/if}}
     {{! End of delete options }}
     {{#if @secret.canReadMetadata}}
