+++ ui/lib/kv/addon/components/page/secret/details.js
@@ -7,6 +7,7 @@ import Component from '@glimmer/component';
 import { action } from '@ember/object';
 import { tracked } from '@glimmer/tracking';
 import { next } from '@ember/runloop';
+import { inject as service } from '@ember/service';
 
 /**
  * @module KvSecretDetails renders the key/value data of a KV secret. 
@@ -26,6 +27,11 @@ import { next } from '@ember/runloop';
 
 export default class KvSecretDetails extends Component {
   @tracked showJsonView = false;
+  @tracked deleteModalOpen = false;
+  @tracked destroyModalOpen = false;
+  @tracked deleteType = ''; // when set this is either delete-latest-version or delete-version.
+  @service flashMessages;
+  @service router;
 
   @action
   toggleJsonView() {
@@ -33,10 +39,86 @@ export default class KvSecretDetails extends Component {
   }
 
   @action
+  toggleModal(btn) {
+    btn === 'delete'
+      ? (this.deleteModalOpen = !this.deleteModalOpen)
+      : (this.destroyModalOpen = !this.destroyModalOpen);
+  }
+
+  @action
   onClose(dropdown) {
     // strange issue where closing dropdown triggers full transition (which redirects to auth screen in production)
     // closing dropdown in next tick of run loop fixes it
     next(() => dropdown.actions.close());
+  }
+
+  @action async handleDelete(mode) {
+    // mode is either: delete, destroy or undelete.
+    const deleteType = mode === 'delete' ? this.deleteType : mode;
+    try {
+      await this.args.secret.destroyRecord({
+        adapterOptions: { deleteType, deleteVersions: this.args.secret.version },
+      });
+      this.flashMessages.success(
+        `Successfully ${this.verbToTense(mode, 'past')} Version ${this.args.secret.version} of ${
+          this.args.path
+        }.`
+      );
+      return this.router.transitionTo('vault.cluster.secrets.backend.kv.secret', {
+        queryParams: { version: this.args.secret.version },
+      });
+    } catch {
+      return `There was an issue ${this.verbToTense(mode, 'gerund')} Version ${this.args.secret.version} of ${
+        this.args.path
+      }.`;
+    }
+  }
+
+  get generateDeleteRadioOptions() {
+    let isDeactivated = false;
+    if (this.args.secret.canReadMetadata) {
+      isDeactivated = this.args.metadata?.currentSecret.isDeactivated;
+    }
+
+    return [
+      {
+        key: 'delete-version',
+        label: 'Delete this version',
+        description: `This deletes Version ${this.args.secret.version} of the secret.`,
+        disabled: !this.args.secret.canDeleteVersion,
+        tooltipMessage: 'You do not have permission to delete a specific version.',
+      },
+      {
+        key: 'delete-latest-version',
+        label: 'Delete latest version',
+        description: 'This deletes the most recent version of the secret.',
+        disabled: !this.args.secret.canDeleteLatestVersion || isDeactivated,
+        tooltipMessage: isDeactivated
+          ? `The latest version of the secret is already ${this.args.metadata.currentSecret.state}.`
+          : 'You do not have permission to delete the latest version of this secret.',
+      },
+    ];
+  }
+
+  verbToTense(verb, tense) {
+    if (tense === 'gerund') {
+      // ending in 'ing ex: delete => deleting || destroy => destroying
+      return (
+        verb
+          .replace(/([^aeiouy])y$/, '$1i')
+          .replace(/([^aeiouy][aeiou])([^aeiouy])$/, '$1$2$2')
+          .replace(/e$/, '') + 'ing'
+      );
+    }
+    if (tense === 'past') {
+      // ending in 'ing ex: delete => deleted || destroy => destroyed
+      return (
+        verb
+          .replace(/([^aeiouy])y$/, '$1i')
+          .replace(/([^aeiouy][aeiou])([^aeiouy])$/, '$1$2$2')
+          .replace(/e$/, '') + 'ed'
+      );
+    }
   }
 
   get emptyState() {
