+++ ui/lib/kv/addon/components/page/secret/metadata/details.hbs
@@ -12,7 +12,18 @@
 
   <:toolbarActions>
     {{#if @metadata.canDeleteMetadata}}
-      <KvDeleteModal @mode="delete metadata" @path={{@secret.path}} @secret={{@secret}} @metadata={{@metadata}} />
+      <KvDeleteModal
+        @mode="delete-metadata"
+        @toggleModal={{this.toggleModal}}
+        @modalOpen={{this.deleteModalOpen}}
+        @buttonText="Delete metadata"
+        @handleDelete={{this.handleDelete}}
+      >
+        <:description>
+          This will permanently delete the metadata and versions of the secret. All version history will be removed. This
+          cannot be undone.
+        </:description>
+      </KvDeleteModal>
     {{/if}}
     {{#if @metadata.canUpdateMetadata}}
       <ToolbarLink @route="secret.metadata.edit" data-test-edit-metadata>Edit metadata</ToolbarLink>
